+++ src/ert/config/gen_kw_config.py
@@ -232,7 +232,7 @@ class GenKwConfig(ParameterConfig):
 
     def load_parameters(
         self, ensemble: Ensemble, realizations: npt.NDArray[np.int_]
-    ) -> npt.NDArray[np.float64]:
+    ) -> npt.NDArray[np.float32]:
         return (
             ensemble.load_parameters(self.name, realizations)
             .drop("realization")
@@ -242,7 +242,7 @@ class GenKwConfig(ParameterConfig):
 
     def create_storage_datasets(
         self,
-        from_data: npt.NDArray[np.float64],
+        from_data: npt.NDArray[np.float32],
         iens_active_index: npt.NDArray[np.int_],
     ) -> Iterator[tuple[int | None, pl.DataFrame]]:
         yield (
